+++ app_test_suite/cluster_providers/cluster_provider.py
@@ -28,8 +28,10 @@ class ClusterInfo:
     managing_provider: "ClusterProvider"
     # cluster might have an optional config file used to create that cluster
     config_file: str
-    # a flag indicating if the App Platform was already initialized
-    app_platform_ready: bool = False
+    # a flag indicating if the bundled CRDs were already applied to this cluster
+    crds_ready: bool = False
+    # a flag indicating if Flux was already deployed to this cluster
+    flux_ready: bool = False
 
 
 class ClusterProvider(ABC):